Fix more warnings, tidy up

+++ src/Veldrid.VirtualReality/OpenVR/OpenVRMirrorTexture.cs
@@ -88,7 +88,7 @@ internal class OpenVRMirrorTexture(OpenVRContext context) : IDisposable
         float sampleUVWidth = (float)sampleWidth / eyeWidth;
         float sampleUVHeight = (float)sampleHeight / eyeHeight;
 
-        float max = (float)Math.Max(sampleUVWidth, sampleUVHeight);
+        float max = Math.Max(sampleUVWidth, sampleUVHeight);
         sampleUVWidth /= max;
         sampleUVHeight /= max;
